编译通过。

+++ PGNode.m
@@ -75,7 +75,7 @@ enum {
 
 + (void)initialize
 {
-	srandom(time(NULL)); // Used by our shuffle sort.
+	srandom((unsigned int)time(NULL)); // Used by our shuffle sort.
 }
 
 #pragma mark -PGNode
